+++ index.html
@@ -40,7 +40,7 @@
             </div>
         </div>
     </div>
-    <script src="js/login.mjs"></script>
+    <script src="js/index.mjs" type="module"></script>
 </body>
 
 </html>
