+++ js/auth/apiLinks.mjs
@@ -1,0 +1,8 @@
+//alle api links skal skrives her og exporteres
+
+export const API_BASE_URL = 'https://api.noroff.dev/api/v1';
+
+export const API_REG_URL = `${API_BASE_URL}/social/auth/register`;
+
+export const API_LOGIN_URL = `${API_BASE_URL}/social/auth/login`;
+
