+++ js/auth/loginUser.mjs
@@ -63,3 +63,4 @@ export function loginUser() {
 const postsUrl = `${API_BASE_URL}/social/posts`;
 */
 //getWithToken(postsUrl);
+
