+++ js/handlers/register.mjs
@@ -1,12 +1,3 @@
-//her skal det bare importeres for å registrere ny user. alt jeg trenger for å registrere useren skal utføres her
-
-const API_BASE_URL = 'https://api.noroff.dev/api/v1';
-//End-points:
-//Register:  /social/auth/register
-//Login:  /social/auth/login
-//Get All posts: /social/posts
-
-//import this function rather?
 /**
  * API call that registers new user
  * @param {string} url 
@@ -27,7 +18,7 @@ const API_BASE_URL = 'https://api.noroff.dev/api/v1';
  * //required due to API
  * ```
  */
-async function regUser(url, userData) {
+export async function regUser(url, userData) {
     try {
         const postData = {
             method: 'POST',
@@ -47,23 +38,3 @@ async function regUser(url, userData) {
         console.log(error);
     }
 }
-
-const regForm = document.querySelector("#regform");
-const regFormAction = regForm.getAttribute('action');
-
-regForm.addEventListener('submit', (e) => {
-    e.preventDefault();
-    const newUserName = document.querySelector("#regusername").value;
-    const newUserEmail = document.querySelector("#regemail").value;
-    const newUserPassword = document.querySelector("#regpassword").value;
-    const userToReg = {
-        name: `${newUserName}`,
-        email: `${newUserEmail}`,
-        password: `${newUserPassword}`,
-    }
-    const regUrl = `${API_BASE_URL}/social/auth/register`;
-    regUser(regUrl, userToReg);
-    regForm.submit;
-})
-
-
