+++ js/index.mjs
@@ -1,0 +1,3 @@
+import { API_REG_URL } from "./auth/apiLinks.mjs";
+
+console.log(API_REG_URL)
